Addressed more concerns in PR 92
